# application/server/openapi.yaml
openapi: 3.0.3
info:
  title: Web Speed Hackathon 2026 API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
  /signup:
    post:
      tags:
        - Auth
      summary: サインアップ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - name
              properties:
                username:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - USERNAME_TAKEN
                      - INVALID_USERNAME
  /signin:
    post:
      tags:
        - Auth
      summary: サインイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: connect.sid
      description: express-session によるセッション Cookie
...
  schemas:
    ProfileImage:
      type: object
      required: [id, alt, width, height]
      properties:
        id:
          type: string
          format: uuid
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
    User:
      type: object
      required: [id, username, name, description, createdAt, profileImage]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        profileImage:
          $ref: "#/components/schemas/ProfileImage"
...